backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-filter==22.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
pip-upgrader==1.4.15
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pylibmc==1.6.3
pymemcache==4.0.0
pyparsing==3.0.9
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
style==1.1.0
terminaltables==3.1.10
tomli==2.0.1
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.14
